WIP: All embeddings have mandatory ref, ref as a flexible trait
